requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Pillow==10.0.1